requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
transformers>=4.35.0
torch>=2.0.0